Implement grid current limiting
When in normal mode and max_grid_current is configured, keep actively managing the grid current, in order to stay below the maximum and protect the installation's main breakers.
If no updates are received on the grid_ie feed for a while, time out and revert to a safe current setting.